+++ lib/space_checker.rb
@@ -5,7 +5,7 @@ module SpaceChecker
     err
   end
 
-  def self.found_spaces(cont)
+  private_class_method def self.found_spaces?(cont)
     spaces_before = /[\s]+(const|let|var|function|class)[\s]+[\w]*/
     spaces_after = /[\s]*(const|let|var|function|class)[\s]{2,}/
     spaced_console = /[\s]+(function|(console.log)[\(][\w]*[\)])[\s]*/
@@ -13,11 +13,11 @@ module SpaceChecker
     spaces_before.match?(cont) || spaces_after.match?(cont) || spaced_console.match?(cont)
   end
 
-  def self.check_spaces(file)
+  private_class_method def self.check_spaces(file)
     # rubocop:disable Lint/UselessAssignment
     state = false
     lines_with_spaces = []
-    file.lines.each { |line| lines_with_spaces << line.number if found_spaces(line.content) }
+    file.lines.each { |line| lines_with_spaces << line.number if found_spaces?(line.content) }
     size = lines_with_spaces.count
 
     state = puts "#{lines_with_spaces}: SpacingError: Redundant Space(s) found".red if size.positive?
