+++ lib/var_naming.rb
@@ -5,17 +5,17 @@ module VarNaming
     err
   end
 
-  def self.bad_var_case(bad_case)
+  private_class_method def self.bad_var_case?(bad_case)
     bad_var_start = /(var|let|const|[\s])[\s]*([[:upper:]]{1,}|\d)+(([\w]+[\s][\w]+)|[\w]+)[\s]*[\=][\s]*[\w]*/
     commented_line = bad_case.match?(%r{^\W+[\/\/]})
 
     bad_var_start.match?(bad_case) && !commented_line
   end
 
-  def self.check_naming(file)
+  private_class_method def self.check_naming(file)
     bad_lines = []
     file.lines.each do |line|
-      bad_lines << line.number if bad_var_case(line.content)
+      bad_lines << line.number if bad_var_case?(line.content)
     end
     size = bad_lines.count
